requirements.txt
Flask
flask_sqlalchemy
pymysql
python-dotenv